Add property_title rewrite

+++ llmApp/migrations/0001_initial.py
@@ -29,7 +29,7 @@ class Migration(migrations.Migration):
             name='PropertyReview',
             fields=[
                 ('id', models.BigAutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
-                ('rating', models.DecimalField(decimal_places=1, max_digits=3)),
+                ('rating', models.FloatField()),
                 ('review', models.TextField()),
                 ('created_at', models.DateTimeField(auto_now_add=True)),
                 ('updated_at', models.DateTimeField(auto_now=True)),
